fp16

+++ config/experiment_config.py
@@ -26,7 +26,7 @@ def read_cfg(mod="non_contrastive"):
     # two options [fixed_value, schedule] schedule recommend from BYOL
     FLAGS.moving_average = "schedule"
     # ['fp16', 'fp32'],  # fp32 is original precision
-    FLAGS.mixprecision = 'fp32'
+    FLAGS.mixprecision = 'fp16'
 
     FLAGS.base_lr = 0.5
